requirements.txt
requests
RPi.GPIO
soco
soundcloud
